python/requirements.txt
pynput
termcolor
python-socketio[client]